// src/app/models/pedido.ts
export class pedidoInv {
    idpedidoinv: number;
    idsucursal: number;
    idprodinv: number;
    cantidad: number;
    existencia: number;
    existenciabodega: number;
    nombreproducto: string;
    fechaentrega: string;
    estado: string;
    iduseradd: number;
    dateadd: string;
    idusermod: number;
    datemod: string;
    serie: string;
}